# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/check-username": {
      "post": {
        "operationId": "AuthController_checkUsername",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckUsernameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Username is available",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "summary": "Check if username is available",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/confirm": {
      "get": {
        "operationId": "AuthController_confirm",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "username",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uid",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User registration confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "summary": "Confirm user registration",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/send-confirmation-email": {
      "post": {
        "description": "This endpoint requires a valid access token for authorization.",
        "operationId": "AuthController_sendConfirmationEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserMailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Confirmation email sent"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send confirmation email",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/send-password-reset-email": {
      "post": {
        "description": "This endpoint requires a valid access token for authorization.",
        "operationId": "AuthController_sendPasswordResetEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserMailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset email sent"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send password reset email",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        },
        "summary": "Login to the system",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "This endpoint requires a valid access token for authorization.",
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user information",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/confirm-token-reset-password": {
      "post": {
        "description": "This endpoint requires a valid access token for authorization.",
        "operationId": "AuthController_confirmTokenResetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token confirmed for password reset",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm token for password reset",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/update": {
      "put": {
        "description": "This endpoint requires a valid access token for authorization.",
        "operationId": "AuthController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user information",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "description": "This endpoint allows users to request a password reset email.",
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserForgotPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset email sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserForgotPasswordResponse"
                }
              }
            }
          }
        },
        "summary": "Request password reset",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh-token": {
      "post": {
        "description": "This endpoint allows users to refresh their access token using a valid refresh token.",
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Access token refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Refresh access token",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Project Management API",
    "description": "Project Management API",
    "version": "1.0.0",
    "contact": {
      "name": "maulanadityaa",
      "url": "https://github.com/maulanadityaa",
      "email": "maulanadityaaa@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3030"
    },
    {
      "url": "https://project-management-api-cyan.vercel.app"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CheckUsernameRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "test_username",
            "description": "Username"
          }
        },
        "required": [
          "username"
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "test_username",
            "description": "Username"
          },
          "email": {
            "type": "string",
            "example": "email@example.com",
            "description": "Email"
          },
          "name": {
            "type": "string",
            "example": "test_name",
            "description": "Name"
          },
          "password": {
            "type": "string",
            "example": "test_password",
            "description": "Password"
          }
        },
        "required": [
          "username",
          "email",
          "name",
          "password"
        ]
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "validJwtToken",
            "description": "Token"
          },
          "isEmailSent": {
            "type": "boolean",
            "example": true,
            "description": "Is Email Sent"
          }
        },
        "required": [
          "token",
          "isEmailSent"
        ]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "test_username",
            "description": "Username"
          },
          "name": {
            "type": "string",
            "example": "test_name",
            "description": "Name"
          },
          "uid": {
            "type": "string",
            "example": "ValidUUIDv4",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "example": "example@mail.com",
            "description": "Email"
          }
        },
        "required": [
          "username",
          "name",
          "uid",
          "email"
        ]
      },
      "UserMailRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "username",
            "description": "Username"
          },
          "uid": {
            "type": "string",
            "example": "ValidUUIDv4",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "example": "example@email.com",
            "description": "Email (optional -- for send reset password only)"
          }
        },
        "required": [
          "username",
          "uid"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "test_username",
            "description": "Username"
          },
          "password": {
            "type": "string",
            "example": "test_password",
            "description": "Password"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "validJwtToken",
            "description": "Token"
          },
          "isConfirmed": {
            "type": "boolean",
            "example": true,
            "description": "Is User Confirmed"
          }
        },
        "required": [
          "token",
          "isConfirmed"
        ]
      },
      "PasswordResetRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "validCode",
            "description": "Code"
          },
          "username": {
            "type": "string",
            "example": "test_username",
            "description": "Username"
          },
          "uid": {
            "type": "string",
            "example": "ValidUUIDv4",
            "description": "User ID"
          }
        },
        "required": [
          "code",
          "username",
          "uid"
        ]
      },
      "UserUpdateRequest": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string",
            "example": "ValidUUIDv4",
            "description": "User ID"
          },
          "name": {
            "type": "string",
            "example": "test_name",
            "description": "Name (optional)"
          },
          "password": {
            "type": "string",
            "example": "test_password",
            "description": "Password (optional)"
          },
          "code": {
            "type": "string",
            "example": "code",
            "description": "Password reset code (optional -- for password reset only)"
          }
        },
        "required": [
          "uid"
        ]
      },
      "UserForgotPasswordRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "example@email.com",
            "description": "Email"
          }
        },
        "required": [
          "email"
        ]
      },
      "UserForgotPasswordResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "validJwtToken",
            "description": "Token"
          },
          "isEmailSent": {
            "type": "boolean",
            "example": true,
            "description": "Is Email Sent"
          }
        },
        "required": [
          "token",
          "isEmailSent"
        ]
      },
...
    }
  }
}